requirements.txt
flask
flask-cors
tensorflow
pillow
opencv-python-headless
gunicorn
